Default output format now RGBA16. Intermediate format now FLOAT RGBA. Added -K option ( error diffuion kernel ). By default, no error diffusion is performed. ( best compression ). IF the final output format has a lower bitrate that the output format, and a diffusion kernel is selected, error diffusion is performed.

+++ rh_texture_packer/Output.hpp
@@ -180,7 +180,7 @@ public:
 	  int hashes = 0;
 
 	  if(args.debug)
-		printf("writing hashmap at 0x%08x (size=%d)\n", (int)ftell(file), spriteMap.size());
+		printf("writing hashmap at 0x%08x (size=%d)\n", (int)ftell(file), (int)spriteMap.size());
 
       while(itor != end) {
